requirements.txt
torch==1.10
torchvision
scikit-learn
numpy
tqdm
pandas
Pillow
tensorflow